requirements.txt
fastapi
pydantic
uvicorn
motor
pydantic[email]
python-jose
py-bcrypt
python-multipart